Откат до поставленных задач в третьем задании. Перенес логику взаимодействия с конструктором пиццы из App в Index
