+++ src/frontend/src/assets/scss/app.scss
@@ -18,13 +18,13 @@
 @import "layout/sign-form.scss";
 
 // blocks
-//@import "blocks/icon.scss";
+@import "blocks/icon.scss";
 @import "blocks/logo.scss";
 @import "blocks/close.scss";
 @import "blocks/input.scss";
 @import "blocks/select.scss";
 @import "blocks/radio.scss";
-//@import "blocks/button.scss";
+@import "blocks/button.scss";
 @import "blocks/counter.scss";
 @import "blocks/title.scss";
 @import "blocks/filling.scss";
