+++ src/frontend/src/common/components/AppButton.vue
@@ -57,6 +57,4 @@ export default {
 };
 </script>
 
-<style lang="scss" scoped>
-@import "~@/assets/scss/blocks/button.scss";
-</style>
+<style lang="scss" scoped></style>
