--- src/frontend/src/modules/cart/componentns/CartForm.vue
@@ -1,51 +1,0 @@
-<template>
-  <div class="cart-form">
-    <label class="cart-form__select">
-      <span class="cart-form__label">Получение заказа:</span>
-
-      <select name="test" class="select">
-        <option value="1">Заберу сам</option>
-        <option value="2">Новый адрес</option>
-        <option value="3">Дом</option>
-      </select>
-    </label>
-
-    <label class="input input--big-label">
-      <span>Контактный телефон:</span>
-      <input type="text" name="tel" placeholder="+7 999-999-99-99" />
-    </label>
-
-    <div class="cart-form__address">
-      <span class="cart-form__label">Новый адрес:</span>
-
-      <div class="cart-form__input">
-        <label class="input">
-          <span>Улица*</span>
-          <input type="text" name="street" />
-        </label>
-      </div>
-
-      <div class="cart-form__input cart-form__input--small">
-        <label class="input">
-          <span>Дом*</span>
-          <input type="text" name="house" />
-        </label>
-      </div>
-
-      <div class="cart-form__input cart-form__input--small">
-        <label class="input">
-          <span>Квартира</span>
-          <input type="text" name="apartment" />
-        </label>
-      </div>
-    </div>
-  </div>
-</template>
-
-<script>
-export default {
-  name: "CartForm",
-};
-</script>
-
-<style scoped></style>
